Create alter_date_cols
When copying date columns from the .csv, the basic 'COPY' code doesn't like inserting the date string into a column with type 'DATE'.  One solution is to copy data into a column of type 'VARCHAR(10)' and then use an 'ALTER' statement to change the data type afterwards.  This code changes the type of all date columns in the table 'basiccompanydata' from 'VARCHAR(10)' to 'DATE'.